feat(v2-tests): tests API integración (54 tests) + fixes previos a Fase 4
- tests/api/: suite completa con 54 tests para auth, student, teacher y admin
- api/routers/auth.py: compatibilidad de placeholder SQL (? vs %s) para SQLite/Postgres
- api/routers/student.py: correct_option se obtiene de DB (no del cliente) por seguridad; validación de curso al matricular
- api/routers/teacher.py: dashboard retorna {students, groups} (tupla desempaquetada)
- api/schemas/student.py: difficulty como float (no int) para compatibilidad
- teacher_service.py + ambos repos: create_group/create_new_group retorna 3-tupla con group_id
- ambos repos: agregar get_item_by_id() para resolver ítem canónico por ID
- frontend/src/pages/Login.tsx: setAuth antes de llamar me() para incluir token en header
- .github/workflows/ci.yml: jobs 6 (test-api) y 7 (build-frontend) agregados

+++ api/schemas/student.py
@@ -27,7 +27,7 @@ class NextQuestionRequest(BaseModel):
 class ItemResponse(BaseModel):
     id: str
     content: str
-    difficulty: int
+    difficulty: float
     topic: str
     options: list[str]
     image_url: str | None = None
